+++ tests/trainer/test_trainer.py
@@ -6,7 +6,6 @@ from argparse import Namespace
 
 import pytest
 import torch
-import yaml
 
 import tests.base.utils as tutils
 from pytorch_lightning import Callback, LightningModule
@@ -17,52 +16,6 @@ from pytorch_lightning.loggers import TensorBoardLogger
 from pytorch_lightning.trainer.logging import TrainerLoggingMixin
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
 from tests.base import EvalModelTemplate
-
-
-def test_auto_hparams(tmpdir):
-    class SubClassEvalModelTemplate(EvalModelTemplate):
-        def __init__(self, subclass_arg=1200):
-            super().__init__()
-
-    class SubSubClassEvalModelTemplate(SubClassEvalModelTemplate):
-        pass
-
-    classes = [SubClassEvalModelTemplate, EvalModelTemplate, SubSubClassEvalModelTemplate]
-
-    for CLASS in classes:
-        # test that the model automatically sets the args passed into init as attrs
-        model = CLASS()
-        assert model.batch_size == 32
-        model = CLASS(batch_size=179)
-        assert model.batch_size == 179
-
-        if isinstance(model, SubClassEvalModelTemplate):
-            assert model.subclass_arg == 1200
-
-        # verify that the checkpoint saved the correct values
-        trainer = Trainer(max_steps=20)
-        trainer.fit(model)
-        raw_checkpoint_path = os.listdir(trainer.checkpoint_callback.dirpath)
-        raw_checkpoint_path = [x for x in raw_checkpoint_path if '.ckpt' in x][0]
-        raw_checkpoint_path = os.path.join(trainer.checkpoint_callback.dirpath, raw_checkpoint_path)
-        raw_checkpoint = torch.load(raw_checkpoint_path)
-        assert 'module_arguments' in raw_checkpoint
-        assert raw_checkpoint['module_arguments']['batch_size'] == 179
-
-        # verify that model loads correctly
-        model = CLASS.load_from_checkpoint(raw_checkpoint_path)
-        assert model.batch_size == 179
-
-        # verify that we can overwrite whatever we want
-        model = CLASS.load_from_checkpoint(raw_checkpoint_path, batch_size=99)
-        assert model.batch_size == 99
-
-
-def test_model_pickle(tmpdir):
-    import pickle
-
-    model = EvalModelTemplate()
-    pickle.dumps(model)
 
 
 def test_dict_namespace_param_save_load(tmpdir):
